# docs/07_api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zedly API v1",
    "version": "0.1.0",
    "description": "Executable API contract for current /api/v1 endpoints.",
    "contact": {
      "name": "Zedly API Team"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/users/provision": {
      "post": {
        "operationId": "usersProvision",
        "tags": [
          "users"
        ],
        "description": "Provision account by hierarchy role with auto-generated login and OTP.",
        "responses": {
          "201": {
            "description": "Account provisioned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeProvisionAccount"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersProvisionRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "usersMeGet",
        "tags": [
          "users"
        ],
        "description": "Get current authenticated user profile.",
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeUser"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "usersMePatch",
        "tags": [
          "users"
        ],
        "description": "Update current authenticated user profile.",
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeUser"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersPatchMeRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/schools/{school_id}/users": {
      "get": {
        "operationId": "schoolUsersList",
        "tags": [
          "users"
        ],
        "description": "List school users with optional filters.",
        "parameters": [
          {
            "name": "school_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeSchoolUsers"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schools/{school_id}/users/{user_id}": {
      "patch": {
        "operationId": "schoolUserPatch",
        "tags": [
          "users"
        ],
        "description": "Update school user status.",
        "parameters": [
          {
            "name": "school_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeUser"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchoolUserPatchRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/classes/{class_id}/invite": {
      "post": {
        "operationId": "classInviteCreate",
        "tags": [
          "classes"
        ],
        "description": "Create class invite code for class distribution flow.",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invite created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessEnvelopeClassInvite"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "example": {
          "code": "string",
          "message": "string"
        }
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "required": [
          "ok",
          "error"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "examples": [
          {
            "ok": false,
            "error": {
              "code": "resource_not_found",
              "message": "Entity not found",
              "details": {
                "resource": "report",
                "resource_id": "rpt_missing"
              }
            }
          }
        ]
      },
...
      "ProvisionAccountData": {
        "type": "object",
        "required": [
          "user_id",
          "role",
          "login",
          "otp_password",
          "status"
        ],
        "properties": {
          "user_id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "otp_password": {
            "type": "string"
          },
          "school_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "district_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "class_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "example": {
          "user_id": "usr_123",
          "role": "student",
          "login": "ali.karimov.9a.47",
          "otp_password": "Xk7#mP2q",
          "school_id": "school_A",
          "district_id": "district_X",
          "class_id": "cls_A_7A",
          "status": "active",
          "message": "Account provisioned"
        }
      },
...
      "TeacherClassRef": {
        "type": "object",
        "required": [
          "class_id",
          "class_name"
        ],
        "properties": {
          "class_id": {
            "type": "string"
          },
          "class_name": {
            "type": "string"
          }
        },
        "example": {
          "class_id": "string",
          "class_name": "string"
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "school_id",
          "role",
          "full_name",
          "login",
          "status",
          "language"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "school_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "teacher_classes": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/TeacherClassRef"
            }
          },
          "login_methods": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "example": {
          "id": "string",
          "school_id": "string",
          "role": "string",
          "full_name": "string",
          "login": "string",
          "status": "string",
          "language": "string"
        }
      },
      "SchoolUsersResponseData": {
        "type": "object",
        "required": [
          "school_id",
          "users",
          "total_in_scope",
          "filtered_total",
          "role",
          "status"
        ],
        "properties": {
          "school_id": {
            "type": "string"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "total_in_scope": {
            "type": "integer",
            "minimum": 0
          },
          "filtered_total": {
            "type": "integer",
            "minimum": 0
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "class_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "search": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "example": {
          "school_id": "string",
          "users": [
            {
              "id": "string",
              "school_id": "string",
              "role": "string",
              "full_name": "string",
              "login": "string",
              "status": "string",
              "language": "string"
            }
          ],
          "total_in_scope": 0,
          "filtered_total": 0,
          "role": "string",
          "status": "string"
        }
      },
      "ClassInviteResponseData": {
        "type": "object",
        "required": [
          "class_id",
          "code",
          "expires_at"
        ],
        "properties": {
          "class_id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "expires_at": {
            "type": "string"
          }
        },
        "example": {
          "class_id": "string",
          "code": "string",
          "expires_at": "string"
        }
      },
...
      "ApiSuccessEnvelopeProvisionAccount": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/ProvisionAccountData"
          }
        }
      },
...
      "ApiSuccessEnvelopeUser": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/User"
          }
        },
        "examples": [
          {
            "ok": true,
            "data": {
              "id": "usr_teacher_001",
              "school_id": "sch_001",
              "role": "teacher",
              "full_name": "Amina Teacher",
              "login": "amina.teacher.7a.12",
              "email": "amina.teacher@school.local",
              "phone": null,
              "status": "active",
              "language": "ru",
              "avatar_url": null,
              "login_methods": [
                "password"
              ],
              "teacher_classes": [
                {
                  "class_id": "class_7a",
                  "class_name": "7A"
                }
              ]
            }
          }
        ]
      },
      "ApiSuccessEnvelopeSchoolUsers": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/SchoolUsersResponseData"
          }
        },
        "examples": [
          {
            "ok": true,
            "data": {
              "school_id": "sch_001",
              "users": [
                {
                  "id": "usr_student_001",
                  "school_id": "sch_001",
                  "role": "student",
                  "full_name": "Student One",
                  "login": "student.one.7a.12",
                  "email": null,
                  "phone": "+998900000001",
                  "status": "active",
                  "language": "ru",
                  "avatar_url": null,
                  "login_methods": [
                    "password"
                  ],
                  "teacher_classes": null
                }
              ],
              "total_in_scope": 120,
              "filtered_total": 1,
              "role": "student",
              "status": "active",
              "class_id": "class_7a",
              "search": "student"
            }
          }
        ]
      },
      "ApiSuccessEnvelopeClassInvite": {
        "type": "object",
        "required": [
          "ok",
          "data"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/ClassInviteResponseData"
          }
        },
        "example": {
          "ok": true,
          "data": {
            "class_id": "string",
            "code": "string",
            "expires_at": "string"
          }
        }
      },
...
      "UsersProvisionRequest": {
        "type": "object",
        "required": [
          "role",
          "full_name"
        ],
        "properties": {
          "role": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "school_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "district_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "class_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "class_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "subject": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "example": {
          "role": "student",
          "full_name": "Ali Karimov",
          "class_id": "cls_A_7A"
        }
      },
      "UsersPatchMeRequest": {
        "type": "object",
        "properties": {
          "full_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "language": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "example": {
          "full_name": "Teacher A Updated",
          "language": "ru"
        }
      },
      "SchoolUserPatchRequest": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "example": {
          "status": "inactive"
        }
      },
...
    }
  }
}